job-finder-web/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.4,<3.2.0  # Pin to avoid dict cache bug

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Browser Automation
playwright>=1.40.0
playwright-stealth>=1.0.0

# AI/LLM
litellm>=1.30.0

# Security
cryptography>=42.0.0
bleach>=6.1.0
itsdangerous>=2.1.2

# Validation
pydantic>=2.5.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.23.0
pytz>=2024.1
psutil>=5.9.0
aiofiles>=23.2.1
